feat: 최초 로그인 상태 처리 작업 중

+++ src/pages/professor/Mypage.js
@@ -20,6 +20,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCircleExclamation, faPencil, faTrash, faUser } from '@fortawesome/free-solid-svg-icons';
 import { checkValidEmail, checkValidPhone } from '../../modules/regex';
 import api from '../../api/api';
+import ChangePassword from '../../components/ChangePassword';
 
 const Mypage = () => {
   const [disabled, setDisabled] = useState(true);
@@ -29,6 +30,8 @@ const Mypage = () => {
   const [address, setAddress] = useState('');
   const [imgFile, setImgFile] = useState(null);
   const [removeImg, setRemoveImg] = useState(false);
+
+  const [openChangePassword, setOpenChangePassword] = useState(false);
 
   const { user } = useSelector(state => state.main);
   const dispatch = useDispatch();
@@ -64,42 +67,46 @@ const Mypage = () => {
       return;
     }
 
-    try {
-      // TODO: API 작업
-      const formData = new FormData();
+    // 최초 로그인 시 비밀번호 수정 후 정보 수정
+    if (user.profile.secretKey !== 'true') {
+      setOpenChangePassword(true);
+    } else {
+      try {
+        const formData = new FormData();
 
-      const payload = {
-        phone,
-        email,
-        address,
-      };
+        const payload = {
+          phone,
+          email,
+          address,
+        };
 
-      console.log(user?.profile.pic);
+        if (imgFile && user?.profile.pic) {
+          await api.delete(`/api/professor`);
+          formData.append('pic', imgFile);
+        } else if (imgFile && !user?.profile.pic) {
+          formData.append('pic', imgFile);
+        } else if (removeImg) {
+          await api.delete(`/api/professor`);
+        }
 
-      if (imgFile && user?.profile.pic) {
-        await api.delete(`/api/professor`);
-        formData.append('pic', imgFile);
-      } else if (imgFile && !user?.profile.pic) {
-        formData.append('pic', imgFile);
-      } else if (removeImg) {
-        await api.delete(`/api/professor`);
+        formData.append('dto', JSON.stringify(payload));
+
+        await api.put(`/api/professor`, formData, {
+          headers: {
+            'Content-Type': 'multipart/form-data',
+          },
+        });
+
+        dispatch(
+          main.setUser({ ...user, profile: { ...user.profile, phone, email, address, pic: img } })
+        );
+        setDisabled(true);
+        alert('프로필이 업데이트 되었습니다.');
+
+        console.log('change password modal open');
+      } catch (err) {
+        console.log(err);
       }
-
-      formData.append('dto', JSON.stringify(payload));
-
-      await api.put(`/api/professor`, formData, {
-        headers: {
-          'Content-Type': 'multipart/form-data',
-        },
-      });
-
-      dispatch(
-        main.setUser({ ...user, profile: { ...user.profile, phone, email, address, pic: img } })
-      );
-      setDisabled(true);
-      alert('프로필이 업데이트 되었습니다.');
-    } catch (err) {
-      console.log(err);
     }
   };
 
@@ -120,7 +127,9 @@ const Mypage = () => {
     setEmail(user?.profile.email);
     setAddress(user?.profile.address);
     setImg(
-      `http://192.168.0.144:5002/imgs/professor/${user?.profile.iprofessor}/${user?.profile.pic}`
+      user?.profile.pic
+        ? `http://192.168.0.144:5002/imgs/professor/${user?.profile.iprofessor}/${user?.profile.pic}`
+        : null
     );
     setRemoveImg(false);
     setDisabled(true);
@@ -147,159 +156,158 @@ const Mypage = () => {
   };
 
   return (
-    <UserLayout>
-      <TopLayout>
-        <NoticeContainer>
-          <span>* 최대 5MB의 이미지 확장자 파일(.jpeg, .png)만 업로드 가능합니다.</span>
-          <span>* 본인 확인이 불가능한 이미지는 사용이 불가능합니다.</span>
-        </NoticeContainer>
-        <ButtonContainer>
-          <Button onClick={disabled ? () => setDisabled(false) : handleEdit}>
-            {disabled ? '수정' : '저장'}
-          </Button>
-          {!disabled && (
-            <Button negative onClick={handleCancel}>
-              취소
+    <>
+      <UserLayout>
+        <TopLayout>
+          <NoticeContainer>
+            <span>* 최대 5MB의 이미지 확장자 파일(.jpeg, .png)만 업로드 가능합니다.</span>
+            <span>* 본인 확인이 불가능한 이미지는 사용이 불가능합니다.</span>
+          </NoticeContainer>
+          <ButtonContainer>
+            <Button onClick={disabled ? () => setDisabled(false) : handleEdit}>
+              {disabled ? '수정' : '저장'}
             </Button>
-          )}
-          <Button>비밀번호 변경</Button>
-        </ButtonContainer>
-      </TopLayout>
-      <MiddleLayout>
-        <ProfileImage>
-          {img ? <img src={img} alt="프로필 이미지" /> : <FontAwesomeIcon icon={faUser} />}
-          {!disabled && (
-            <ModifyImage>
-              <label>
-                <FontAwesomeIcon icon={faPencil} />
-                {img ? '수정' : '등록'}
-                <input type="file" accept="image/jpeg, image/png" onChange={handleUpdateImage} />
-              </label>
-              {img && (
-                <CancelModifyButton onClick={handleRemoveImage}>
-                  <FontAwesomeIcon icon={faTrash} />
-                  삭제
-                </CancelModifyButton>
-              )}
-            </ModifyImage>
-          )}
-        </ProfileImage>
-        <LectureTableLayout>
-          <div className="lecture-table-header">
-            <div>강의명</div>
-            <div>강의 기간</div>
-            <div>강의 시간</div>
-          </div>
-          <div className="lecture-table-body">
-            {user?.lectureList?.length === 0 && (
-              <div className="lecture-table-no-content">
-                <FontAwesomeIcon icon={faCircleExclamation} />
-                <span>강의 중인 강의가 없습니다.</span>
-              </div>
+            {!disabled && (
+              <Button negative onClick={handleCancel}>
+                취소
+              </Button>
             )}
-            {user?.lectureList?.map((item, index) => (
-              <div key={index} className="lecture-table-content">
-                <div>{item.lectureName}</div>
-                <div>{`${item.lectureStrDate} ~ ${item.lectureEndDate}`}</div>
-                <div>{`${item.lectureStrTime} ~ ${item.lectureEndTime}`}</div>
-              </div>
-            ))}
-            {user?.lectureList?.length <= 7 &&
-              Array(7 - (user?.lectureList?.length ?? 0))
-                .fill()
-                .map((_, index) => (
-                  <div key={index} className="lecture-table-content">
-                    <div></div>
-                    <div></div>
-                    <div></div>
-                  </div>
-                ))}
-          </div>
-        </LectureTableLayout>
-      </MiddleLayout>
-      <FormTable>
-        <Row col={2}>
-          <div>이름</div>
-          <div>{user?.profile.name}</div>
-          <div>성별</div>
-          <div>
-            {(user?.profile.gender === 'M' && '남') || (user?.profile.gender === 'F' && '여')}
-          </div>
-        </Row>
-        <Row col={2}>
-          <div>생년월일</div>
-          <div>{user?.profile.birthdate}</div>
-          <div>전공</div>
-          <div>{user?.profile.majorName}</div>
-        </Row>
-        <Row col={2}>
-          <div>등록일</div>
-          <div>{user?.profile.createdAt?.split('T')[0]}</div>
-          <div>퇴직 여부</div>
-          <div>
-            {(user?.profile.delYn === 0 && '재직 중') || (user?.profile.delYn === 1 && '퇴직')}
-          </div>
-        </Row>
-        <Row col={2}>
-          <div>
-            {!disabled && <FontAwesomeIcon icon={faPencil} />}
-            휴대전화
-          </div>
-          <div>
-            <Input
-              type="text"
-              isForm={true}
-              placeholder="010-0000-0000"
-              reset={setPhone}
-              maxLength={13}
-              value={phone || ''}
-              setValue={handlePhoneChange}
-              disabled={disabled}
-            />
-          </div>
-          <div>
-            E-mail
-            {!disabled && <FontAwesomeIcon icon={faPencil} />}
-          </div>
-          <div>
-            {user?.profile.email ? (
+            <Button>비밀번호 변경</Button>
+          </ButtonContainer>
+        </TopLayout>
+        <MiddleLayout>
+          <ProfileImage>
+            {img ? <img src={img} alt="프로필 이미지" /> : <FontAwesomeIcon icon={faUser} />}
+            {!disabled && (
+              <ModifyImage>
+                <label>
+                  <FontAwesomeIcon icon={faPencil} />
+                  {img ? '수정' : '등록'}
+                  <input type="file" accept="image/jpeg, image/png" onChange={handleUpdateImage} />
+                </label>
+                {img && (
+                  <CancelModifyButton onClick={handleRemoveImage}>
+                    <FontAwesomeIcon icon={faTrash} />
+                    삭제
+                  </CancelModifyButton>
+                )}
+              </ModifyImage>
+            )}
+          </ProfileImage>
+          <LectureTableLayout>
+            <div className="lecture-table-header">
+              <div>강의명</div>
+              <div>강의 기간</div>
+              <div>강의 시간</div>
+            </div>
+            <div className="lecture-table-body">
+              {user?.lectureList?.length === 0 && (
+                <div className="lecture-table-no-content">
+                  <FontAwesomeIcon icon={faCircleExclamation} />
+                  <span>강의 중인 강의가 없습니다.</span>
+                </div>
+              )}
+              {user?.lectureList?.map((item, index) => (
+                <div key={index} className="lecture-table-content">
+                  <div>{item.lectureName}</div>
+                  <div>{`${item.lectureStrDate} ~ ${item.lectureEndDate}`}</div>
+                  <div>{`${item.lectureStrTime} ~ ${item.lectureEndTime}`}</div>
+                </div>
+              ))}
+              {user?.lectureList?.length <= 7 &&
+                Array(7 - (user?.lectureList?.length ?? 0))
+                  .fill()
+                  .map((_, index) => (
+                    <div key={index} className="lecture-table-content">
+                      <div></div>
+                      <div></div>
+                      <div></div>
+                    </div>
+                  ))}
+            </div>
+          </LectureTableLayout>
+        </MiddleLayout>
+        <FormTable>
+          <Row col={2}>
+            <div>이름</div>
+            <div>{user?.profile.name}</div>
+            <div>성별</div>
+            <div>
+              {(user?.profile.gender === 'M' && '남') || (user?.profile.gender === 'F' && '여')}
+            </div>
+          </Row>
+          <Row col={2}>
+            <div>생년월일</div>
+            <div>{user?.profile.birthdate}</div>
+            <div>전공</div>
+            <div>{user?.profile.majorName}</div>
+          </Row>
+          <Row col={2}>
+            <div>등록일</div>
+            <div>{user?.profile.createdAt?.split('T')[0]}</div>
+            <div>퇴직 여부</div>
+            <div>
+              {(user?.profile.delYn === 0 && '재직 중') || (user?.profile.delYn === 1 && '퇴직')}
+            </div>
+          </Row>
+          <Row col={2}>
+            <div>
+              {!disabled && <FontAwesomeIcon icon={faPencil} />}
+              휴대전화
+            </div>
+            <div>
               <Input
                 type="text"
                 isForm={true}
-                placeholder="smartgrade@green.ac.kr"
+                placeholder="010-0000-0000"
+                reset={setPhone}
+                maxLength={13}
+                value={phone || ''}
+                setValue={handlePhoneChange}
+                disabled={disabled}
+              />
+            </div>
+            <div>
+              E-mail
+              {!disabled && <FontAwesomeIcon icon={faPencil} />}
+            </div>
+            <div>
+              <Input
+                type="text"
+                isForm={true}
+                placeholder={
+                  disabled && !user?.profile.email ? '(정보 수정 필요)' : 'smartgrade@green.ac.kr'
+                }
                 reset={setEmail}
                 value={email || ''}
                 setValue={handleEmailChange}
                 disabled={disabled}
               />
-            ) : (
-              <span>(정보 수정 필요)</span>
-            )}
-          </div>
-        </Row>
-        <Row>
-          <div>
-            주소
-            {!disabled && <FontAwesomeIcon icon={faPencil} />}
-          </div>
-          <div>
-            {user?.profile.address ? (
+            </div>
+          </Row>
+          <Row>
+            <div>
+              주소
+              {!disabled && <FontAwesomeIcon icon={faPencil} />}
+            </div>
+            <div>
               <Input
                 type="text"
                 isForm={true}
-                placeholder="주소를 입력하세요."
+                placeholder={
+                  disabled && !user?.profile.address ? '(정보 수정 필요)' : '주소를 입력하세요.'
+                }
                 reset={setAddress}
                 value={address || ''}
                 setValue={handleAddressChange}
                 disabled={disabled}
               />
-            ) : (
-              <span>(정보 수정 필요)</span>
-            )}
-          </div>
-        </Row>
-      </FormTable>
-    </UserLayout>
+            </div>
+          </Row>
+        </FormTable>
+      </UserLayout>
+      {openChangePassword && <ChangePassword setOpen={setOpenChangePassword} />}
+    </>
   );
 };
